+++ Model/QuoteHandler.php
@@ -13,6 +13,8 @@
 namespace Magestat\SplitOrder\Model;
 
 use Magento\Checkout\Model\Session as CheckoutSession;
+use Magento\Quote\Api\CartManagementInterface;
+use Magento\Customer\Api\Data\GroupInterface;
 use Magestat\SplitOrder\Api\QuoteHandlerInterface;
 use Magestat\SplitOrder\Helper\Data as HelperData;
 
@@ -45,23 +47,44 @@ class QuoteHandler implements QuoteHandlerInterface
      */
     public function normalizeQuotes($quote)
     {
+        // if module is active.
+        if (!$this->helperData->isActive()) {
+            return false;
+        }
+        $attributes = $this->helperData->getAttributes();
+        if (empty($attributes)) {
+            return false;
+        }
+
         $groups = [];
         foreach ($quote->getAllVisibleItems() as $item) {
-            $groups[$item->getProduct()->getSku()][] = $item;
+            $product = $item->getProduct();
+
+            $attribute = $this->getProductAttributes($product, $attributes);
+            if (empty($attribute)) {
+                // Kill function if attribute is null.
+                return false;
+            }
+            $groups[$attribute][] = $item;
         }
-        return $groups;
+        // If order have more than one item.
+        if (count($groups) > 1) {
+            return $groups;
+        }
+        return false;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function isSplittable($quotes)
+    public function getProductAttributes($product, $attributes)
     {
-        // If module is active and order have more than one item.
-        if ($this->helperData->isActive() && count($quotes) > 1) {
-            return true;
-        }
-        return false;
+        $attribute = $product->getResource()
+            ->getAttribute($attributes)
+            ->getFrontend()
+            ->getValue($product);
+
+        return $attribute;
     }
 
     /**
@@ -93,11 +116,11 @@ class QuoteHandler implements QuoteHandlerInterface
         $split->setCustomer($quote->getCustomer());
         $split->setCustomerIsGuest($quote->getCustomerIsGuest());
 
-        if ($quote->getCheckoutMethod() === \Magento\Quote\Api\CartManagementInterface::METHOD_GUEST) {
+        if ($quote->getCheckoutMethod() === CartManagementInterface::METHOD_GUEST) {
             $split->setCustomerId(null);
             $split->setCustomerEmail($quote->getBillingAddress()->getEmail());
             $split->setCustomerIsGuest(true);
-            $split->setCustomerGroupId(\Magento\Customer\Api\Data\GroupInterface::NOT_LOGGED_IN_ID);
+            $split->setCustomerGroupId(GroupInterface::NOT_LOGGED_IN_ID);
         }
         return $this;
     }
@@ -105,12 +128,11 @@ class QuoteHandler implements QuoteHandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function populateQuote($split, $item, $addresses, $paymentMethod)
+    public function populateQuote($quotes, $split, $item, $addresses, $payment)
     {
-        $this->recollectTotal($item, $split, $addresses);
-
+        $this->recollectTotal($quotes, $item, $split, $addresses);
         // Set payment method.
-        $this->setPaymentMethod($split, $addresses['payment'], $paymentMethod);
+        $this->setPaymentMethod($split, $addresses['payment'], $payment);
 
         return $this;
     }
@@ -118,7 +140,7 @@ class QuoteHandler implements QuoteHandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function recollectTotal($item, $quote, $addresses, $shippingAmount = 0)
+    public function recollectTotal($quotes, $item, $quote, $addresses)
     {
         // Retrieve values.
         $tax      = $item->getData('tax_amount');
@@ -131,14 +153,12 @@ class QuoteHandler implements QuoteHandlerInterface
         $quote->getShippingAddress()->setData($addresses['shipping']);
 
         // Add shipping amount if product is not virual.
-        if ($quote->hasVirtualItems() === false) {
-            $shippingAmount = $quote->getShippingAddress()->getShippingAmount();
-        }
+        $shipping = $this->shippingAmount($quotes, $quote);
 
         // Recollect totals into the quote.
         foreach ($quote->getAllAddresses() as $address) {
             // Build grand total.
-            $grandTotal = (($finalPrice + $shippingAmount + $tax) - $discount);
+            $grandTotal = (($finalPrice + $shipping + $tax) - $discount);
 
             $address->setBaseSubtotal($finalPrice);
             $address->setSubtotal($finalPrice);
@@ -150,7 +170,39 @@ class QuoteHandler implements QuoteHandlerInterface
         }
         return $this;
     }
- 
+
+    /**
+     * {@inheritdoc}
+     */
+    public function shippingAmount($quotes, $quote, $total = 0)
+    {
+        // Add shipping amount if product is not virual.
+        if ($quote->hasVirtualItems() === true) {
+            return $total;
+        }
+        $shippingTotals = $quote->getShippingAddress()->getShippingAmount();
+
+        // If not, set shipping to one order only.
+        if (!$this->helperData->getShippingSplit()) {
+            static $process = 1;
+
+            if ($process > 1) {
+                // Set zero price to next orders.
+                $quote->getShippingAddress()->setShippingAmount($total);
+                return $total;
+            }
+            $process ++;
+
+            return $shippingTotals;
+        }
+        if ($shippingTotals > 0) {
+            // Divide shippinto to each order.
+            $total = (float) ($shippingTotals / count($quotes));
+            $quote->getShippingAddress()->setShippingAmount($total);
+        }
+        return $total;
+    }
+
     /**
      * {@inheritdoc}
      */