+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2018-08-21
+### Added
+- Option to select a product attribute to define order split.
+- Capability to split shipping totals into all orders or only one.
+
+### Fixed
+- Split shipping totals into all orders or only one.
+- Set shipping to one order only.
+
+### Changed
+- Improved module architecture.
+- Stable order submission.
+
+
 ## [1.0.1] - 2018-07-20
 ### Added
 - CMS module settings (Enable/Disable)